silence warning about %lu

+++ libhsm/src/hsmtest.c
@@ -86,7 +86,8 @@ hsm_test_random()
     uint32_t r32;
     uint64_t r64;
 
-    printf("Generating %lu bytes of random data... ", sizeof(rnd_buf));
+    printf("Generating %lu bytes of random data... ",
+        (unsigned long) sizeof(rnd_buf));
     result = hsm_random_buffer(ctx, rnd_buf, sizeof(rnd_buf));
     if (result) {
         printf("Failed, error: %d\n", result);
